badgr_text_service/requirements.txt
fastapi
uvicorn[standard]
pymupdf
pdfplumber
python-docx
